Add Lasso example with two solvers, CD and FISTA
CD serves a coordinate descent approach with a closed form solution that
involves less work than FISTA. I think there's something missing for this to
fully work.

FISTA is a generic approach for composite minimization problems where we only
need to plug in the proximal operator of the `g` function.